Old/requirements.txt
kubernetes>=24.2.0
openai>=0.27.0
spacy>=3.5.0
tabulate>=0.9.0
termcolor>=2.2.0
flask>=2.0.0
click>=8.0.0
pyyaml>=6.0
rich>=12.0.0